Revert "Fix/remove chatbot streaming keep thumbdown"
