--- src/courseware/course/AIChatbot-streaming/AIChabotAssets.jsx
@@ -1,238 +1,0 @@
-export const svgChatGPT = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-  >
-    <rect width="32" height="32" rx="2" fill="#007AE6" />
-    <path
-      d="M25.749 13.923a5.7 5.7 0 0 0-.49-4.685 5.767 5.767 0 0 0-6.213-2.767 5.703 5.703 0 0 0-4.302-1.918 5.768 5.768 0 0 0-5.502 3.994 5.705 5.705 0 0 0-3.814 2.766 5.77 5.77 0 0 0 .71 6.764 5.704 5.704 0 0 0 .49 4.685 5.769 5.769 0 0 0 6.212 2.767 5.705 5.705 0 0 0 4.302 1.918 5.768 5.768 0 0 0 5.504-3.996 5.706 5.706 0 0 0 3.814-2.767 5.768 5.768 0 0 0-.711-6.761zM17.144 25.95a4.278 4.278 0 0 1-2.747-.993c.035-.019.096-.052.136-.076l4.558-2.634a.74.74 0 0 0 .375-.648v-6.427l1.926 1.113a.07.07 0 0 1 .038.053v5.322a4.296 4.296 0 0 1-4.286 4.29zm-9.218-3.937a4.276 4.276 0 0 1-.512-2.875c.034.02.093.057.135.081l4.559 2.633a.742.742 0 0 0 .748 0l5.566-3.213v2.225a.069.069 0 0 1-.028.059l-4.608 2.66a4.295 4.295 0 0 1-5.86-1.57zm-1.2-9.951A4.275 4.275 0 0 1 8.96 10.18l-.003.157v5.266a.741.741 0 0 0 .374.648l5.566 3.213-1.927 1.113a.069.069 0 0 1-.065.006L8.297 17.92a4.295 4.295 0 0 1-1.57-5.86zm15.83 3.684-5.565-3.214 1.927-1.112a.069.069 0 0 1 .065-.006l4.608 2.66a4.292 4.292 0 0 1-.663 7.743v-5.423a.74.74 0 0 0-.372-.648zm1.918-2.887a6.286 6.286 0 0 0-.135-.08l-4.559-2.634a.743.743 0 0 0-.748 0l-5.566 3.214v-2.225a.07.07 0 0 1 .028-.06l4.608-2.658a4.291 4.291 0 0 1 6.372 4.443zM12.42 16.825l-1.928-1.112a.068.068 0 0 1-.037-.053v-5.322a4.291 4.291 0 0 1 7.036-3.295 3.944 3.944 0 0 0-.135.077l-4.559 2.633a.74.74 0 0 0-.374.648l-.003 6.424zm1.046-2.256 2.479-1.432 2.479 1.43v2.863l-2.479 1.43-2.479-1.43v-2.861z"
-      fill="#fff"
-    />
-  </svg>
-);
-
-export const svgReset = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="16"
-    height="17"
-    viewBox="0 0 16 17"
-    fill="none"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M5.05919 3.46207C6.17433 2.8109 7.474 2.54811 8.75452 2.71488C10.035 2.88166 11.2241 3.46858 12.1352 4.38366C13.0463 5.29874 13.6281 6.49027 13.7894 7.7715C13.8239 8.04548 13.6297 8.29554 13.3557 8.33002C13.0818 8.36451 12.8317 8.17036 12.7972 7.89637C12.6636 6.83476 12.1815 5.84747 11.4266 5.08925C10.6716 4.33102 9.6864 3.8447 8.62537 3.70651C7.56434 3.56832 6.48744 3.78606 5.56345 4.32562C4.86218 4.73512 4.27907 5.31242 3.86336 6.0006H6.0013C6.27744 6.0006 6.5013 6.22446 6.5013 6.5006C6.5013 6.77675 6.27744 7.0006 6.0013 7.0006H2.66797C2.39183 7.0006 2.16797 6.77675 2.16797 6.5006V3.16727C2.16797 2.89113 2.39183 2.66727 2.66797 2.66727C2.94411 2.66727 3.16797 2.89113 3.16797 3.16727V5.23247C3.65486 4.51232 4.29959 3.90563 5.05919 3.46207ZM2.64651 8.67123C2.92046 8.63655 3.17066 8.83052 3.20534 9.10447C3.33969 10.1657 3.82214 11.1524 4.57718 11.9102C5.33221 12.6679 6.3172 13.1539 7.37794 13.292C8.43868 13.4302 9.51529 13.2127 10.4392 12.6736C11.1402 12.2646 11.7232 11.688 12.1392 11.0006H10.0013C9.72516 11.0006 9.5013 10.7767 9.5013 10.5006C9.5013 10.2245 9.72516 10.0006 10.0013 10.0006H13.3346C13.6108 10.0006 13.8346 10.2245 13.8346 10.5006V13.8339C13.8346 14.1101 13.6108 14.3339 13.3346 14.3339C13.0585 14.3339 12.8346 14.1101 12.8346 13.8339V11.7684C12.3475 12.488 11.7027 13.0942 10.9432 13.5373C9.82817 14.1879 8.52891 14.4504 7.2488 14.2837C5.96868 14.117 4.77999 13.5305 3.86881 12.616C2.95762 11.7016 2.37539 10.5108 2.21326 9.23007C2.17858 8.95611 2.37255 8.70591 2.64651 8.67123Z"
-      fill="black"
-    />
-  </svg>
-);
-
-export const svgUpVote = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="16"
-    height="16"
-    viewBox="0 0 16 16"
-    fill="none"
-  >
-    <path
-      d="m4.661 7.332 2.667-6a2 2 0 0 1 2 2v2.667h3.774a1.333 1.333 0 0 1 1.333 1.533l-.92 6a1.333 1.333 0 0 1-1.334 1.133h-7.52m0-7.333v7.333m0-7.333h-2a1.333 1.333 0 0 0-1.333 1.333v4.667a1.333 1.333 0 0 0 1.333 1.333h2"
-      stroke="#7B7B7B"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-);
-
-export const svgCopy = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="16"
-    height="16"
-    viewBox="0 0 16 16"
-    fill="none"
-  >
-    <path
-      d="M9.875 1h-3.75c-.345 0-.625.313-.625.7v1.4c0 .386.28.7.625.7h3.75c.345 0 .625-.314.625-.7V1.7c0-.387-.28-.7-.625-.7z"
-      stroke="#7B7B7B"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-    <path
-      d="M10.5 2.4h1.25c.332 0 .65.148.884.41.234.263.366.619.366.99v9.8c0 .37-.132.727-.366.99-.235.262-.553.41-.884.41h-7.5c-.332 0-.65-.148-.884-.41A1.49 1.49 0 0 1 3 13.6V3.8c0-.371.132-.727.366-.99.235-.262.552-.41.884-.41H5.5"
-      stroke="#7B7B7B"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-);
-
-export const svgCopied = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="15"
-    height="15"
-    viewBox="0 0 15 15"
-    fill="none"
-  >
-    <path
-      d="M6.625 8.625 5.281 7.281a.593.593 0 0 0-.437-.172.593.593 0 0 0-.438.172.593.593 0 0 0-.172.438c0 .177.058.323.172.437l1.782 1.781a.6.6 0 0 0 .437.188.6.6 0 0 0 .438-.188l3.53-3.53a.593.593 0 0 0 .173-.438.593.593 0 0 0-.172-.438.593.593 0 0 0-.438-.172.593.593 0 0 0-.437.172L6.625 8.625zM7.5 13.75a6.086 6.086 0 0 1-2.438-.492 6.311 6.311 0 0 1-1.984-1.336 6.312 6.312 0 0 1-1.336-1.985A6.086 6.086 0 0 1 1.25 7.5c0-.865.164-1.677.492-2.438a6.311 6.311 0 0 1 1.336-1.984 6.311 6.311 0 0 1 1.985-1.336A6.086 6.086 0 0 1 7.5 1.25c.865 0 1.677.164 2.438.492.76.328 1.421.774 1.984 1.336a6.311 6.311 0 0 1 1.336 1.985c.328.76.492 1.572.492 2.437 0 .865-.164 1.677-.492 2.438a6.311 6.311 0 0 1-1.336 1.984 6.311 6.311 0 0 1-1.985 1.336 6.086 6.086 0 0 1-2.437.492z"
-      fill="#007AE6"
-    />
-  </svg>
-);
-
-export const svgSubmit = (
-  <svg
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <rect
-      x="24"
-      width="24"
-      height="24"
-      rx="4"
-      transform="rotate(90 24 0)"
-      fill="#7B7B7B"
-    />
-    <path
-      d="M11.2137 8.61958L7.33588 12.3949C7.1776 12.549 6.99294 12.6228 6.7819 12.6164C6.57086 12.61 6.3862 12.5297 6.22793 12.3756C6.08283 12.2215 6.00699 12.0417 6.0004 11.8363C5.9938 11.6308 6.06964 11.451 6.22793 11.297L11.4512 6.21188C11.5303 6.13483 11.616 6.08026 11.7084 6.04815C11.8007 6.01605 11.8996 6 12.0051 6C12.1107 6 12.2096 6.01605 12.3019 6.04815C12.3942 6.08026 12.48 6.13483 12.5591 6.21188L17.7824 11.297C17.9275 11.4382 18 11.6148 18 11.8266C18 12.0385 17.9275 12.2215 17.7824 12.3756C17.6241 12.5297 17.4361 12.6067 17.2185 12.6067C17.0009 12.6067 16.8129 12.5297 16.6546 12.3756L12.7965 8.61958L12.7965 17.2295C12.7965 17.4478 12.7207 17.6308 12.569 17.7785C12.4173 17.9262 12.2294 18 12.0051 18C11.7809 18 11.593 17.9262 11.4413 17.7785C11.2896 17.6308 11.2137 17.4478 11.2137 17.2295L11.2137 8.61958Z"
-      fill="white"
-    />
-  </svg>
-);
-
-export const svgSubmitActive = (
-  <svg
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <rect
-      x="24"
-      width="24"
-      height="24"
-      rx="4"
-      transform="rotate(90 24 0)"
-      fill="#007AE6"
-    />
-    <path
-      d="M11.2137 8.61958L7.33588 12.3949C7.1776 12.549 6.99294 12.6228 6.7819 12.6164C6.57086 12.61 6.3862 12.5297 6.22793 12.3756C6.08283 12.2215 6.00699 12.0417 6.0004 11.8363C5.9938 11.6308 6.06964 11.451 6.22793 11.297L11.4512 6.21188C11.5303 6.13483 11.616 6.08026 11.7084 6.04815C11.8007 6.01605 11.8996 6 12.0051 6C12.1107 6 12.2096 6.01605 12.3019 6.04815C12.3942 6.08026 12.48 6.13483 12.5591 6.21188L17.7824 11.297C17.9275 11.4382 18 11.6148 18 11.8266C18 12.0385 17.9275 12.2215 17.7824 12.3756C17.6241 12.5297 17.4361 12.6067 17.2185 12.6067C17.0009 12.6067 16.8129 12.5297 16.6546 12.3756L12.7965 8.61958L12.7965 17.2295C12.7965 17.4478 12.7207 17.6308 12.569 17.7785C12.4173 17.9262 12.2294 18 12.0051 18C11.7809 18 11.593 17.9262 11.4413 17.7785C11.2896 17.6308 11.2137 17.4478 11.2137 17.2295L11.2137 8.61958Z"
-      fill="white"
-    />
-  </svg>
-);
-
-export const svgDontKnow = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    fill="none"
-    viewBox="0 0 24 24"
-    strokeWidth={1.5}
-    stroke="currentColor"
-    className="w-6 h-6"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      d="M6 12 3.269 3.125A59.769 59.769 0 0 1 21.485 12 59.768 59.768 0 0 1 3.27 20.875L5.999 12Zm0 0h7.5"
-    />
-  </svg>
-);
-
-export const svgRetry = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    fill="none"
-    viewBox="0 0 24 24"
-    strokeWidth={1.5}
-    stroke="currentColor"
-    dataSlot="icon"
-    className="w-6 h-6"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      d="M16.023 9.348h4.992v-.001M2.985 19.644v-4.992m0 0h4.992m-4.993 0 3.181 3.183a8.25 8.25 0 0 0 13.803-3.7M4.031 9.865a8.25 8.25 0 0 1 13.803-3.7l3.181 3.182m0-4.991v4.99"
-    />
-  </svg>
-);
-
-// header
-export const svgPlus = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    fill="none"
-    viewBox="0 0 24 24"
-    strokeWidth={1.5}
-    stroke="currentColor"
-    className="w-6 h-6"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      d="M12 4.5v15m7.5-7.5h-15"
-    />
-  </svg>
-);
-
-export const svgHamburger = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    fill="none"
-    viewBox="0 0 24 24"
-    strokeWidth={1.5}
-    stroke="currentColor"
-    className="w-6 h-6"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      d="M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5"
-    />
-  </svg>
-);
-
-export const svgMessage = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    fill="none"
-    viewBox="0 0 24 24"
-    strokeWidth={1.5}
-    stroke="currentColor"
-    className="w-6 h-6"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      d="M2.25 12.76c0 1.6 1.123 2.994 2.707 3.227 1.068.157 2.148.279 3.238.364.466.037.893.281 1.153.671L12 21l2.652-3.978c.26-.39.687-.634 1.153-.67 1.09-.086 2.17-.208 3.238-.365 1.584-.233 2.707-1.626 2.707-3.228V6.741c0-1.602-1.123-2.995-2.707-3.228A48.394 48.394 0 0 0 12 3c-2.392 0-4.744.175-7.043.513C3.373 3.746 2.25 5.14 2.25 6.741v6.018Z"
-    />
-  </svg>
-);
-
-export const svgClose = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="20"
-    height="20"
-    viewBox="0 0 20 20"
-    fill="none"
-  >
-    <path
-      d="m15.834 5.343-1.175-1.175L10 8.826 5.342 4.168 4.167 5.343l4.658 4.658-4.658 4.659 1.175 1.175L10 11.176l4.659 4.659 1.175-1.175L11.175 10l4.659-4.658z"
-      fill="#2D2F31"
-    />
-  </svg>
-);
