+++ src/courseware/course/AIChatbot/AIChatbot.scss
@@ -1,10 +1,6 @@
 .chatbot-container {
   --border-radius: 6px;
 
-  @media screen and (max-width: 992px) {
-    padding: 0 0 0 1rem;
-  }
-
   * {
     box-sizing: border-box;
     margin: 0;
@@ -65,12 +61,17 @@
   width: 100%;
   display: flex;
   flex-direction: column-reverse;
-
   /* width */
+
   &::-webkit-scrollbar {
     width: 0.5rem;
   }
 
+  @media screen and (max-width: 992px) {
+    &::-webkit-scrollbar {
+      display: none;
+    }
+  }
   /* Track */
   &::-webkit-scrollbar-track {
     // background: #f1f1f1;
@@ -78,7 +79,7 @@
 
   /* Handle */
   &::-webkit-scrollbar-thumb {
-    background: #888;
+    background: var(--color-disable);
     border-radius: 999px;
   }
 
@@ -100,6 +101,7 @@
   }
 
   .chatbot-intro-text {
+    width: 17.5rem;
     text-align: center;
     color: var(--color-text-default);
     font-size: 0.875rem;
