+++ src/courseware/course/AIChatbot/AIChatbotHeader.scss
