+++ src/courseware/course/AIChatbot/SessionList.jsx
@@ -3,6 +3,9 @@ import { useSelector, useDispatch } from "react-redux";
 import { useEffect } from "react";
 import { fetchSessions } from "./slice";
 
+import ChatbotListError from "./ChatbotListError";
+
+
 function SessionList({ onSelectSession, mode }) {
   const dispatch = useDispatch();
 
@@ -20,20 +23,9 @@ function SessionList({ onSelectSession, mode }) {
   if (mode === "chat") classes += " d-none";
   return (
     <div className={classes}>
-      {session.error && (
-        <div>
-          <p className="text-center">:( Something wrong happened.</p>
-          <p
-            style={{
-              fontSize: ".75rem",
-              color: "grey",
-              textAlign: "center",
-            }}
-          >
-            {session.error}
-          </p>
-        </div>
-      )}
+
+      {session.error && <ChatbotListError error={session.error} />}
+
 
       {!session.error && (
         <ul className="chatbot-messages-list">
