+++ src/courseware/course/AIChatbot/slice.js
@@ -1,12 +1,16 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import {
-  askChatbot as askChatbotAPI,
+
+  createQueryItemToDB,
+  updateQueryItemToDB,
   fetchSessions as fetchSessionsAPI,
   fetchQueries as fetchQueriesAPI,
-  retryAskChatbot as retryAskChatbotAPI,
   voteChatbotResponse as voteChatbotResponseAPI,
   giveFeedbackChatbot as giveFeedbackChatbotAPI,
+  saveChabotErrorToDB,
 } from "./AIChatbotAPI";
+import { sendMessageToChatbot } from "../../../connection/chatbot";
+
 import * as uid from "uuid";
 
 const LIMIT = 5;
@@ -19,11 +23,12 @@ export const fetchQueries = createAsyncThunk(
       const { session, query } = thunkAPI.getState().chatbot;
       const response = await fetchQueriesAPI(
         session.id,
-        query.items.length,
+
+        query.items.filter((item) => item.session_id === session.id).length,
         LIMIT
       );
 
-      console.log("chatbot queries response", response);
+
       return response.data;
     } catch (error) {
       console.error(error);
@@ -34,17 +39,90 @@ export const fetchQueries = createAsyncThunk(
 
 export const askChatbot = createAsyncThunk(
   "chatbot/askChatbot",
+
   async (askInput, thunkAPI) => {
     try {
-      const { ask, session } = thunkAPI.getState().chatbot;
+      const { session, courseId } = thunkAPI.getState().chatbot;
 
-      const response = await askChatbotAPI(
-        askInput,
-        session.id,
-        thunkAPI.requestId
-      );
+      const response = await createQueryItemToDB({
+        query_msg: askInput,
+        session_id: session.id,
+        hash: thunkAPI.requestId,
+        status: "idle",
+      });
+
+      sendMessageToChatbot(askInput, session.id, courseId);
+
       return response.data;
     } catch (error) {
+      console.error(error);
+      return thunkAPI.rejectWithValue(_composeErrorMessage(error));
+    }
+  }
+);
+
+
+export const retryAskChatbot = createAsyncThunk(
+  "chatbot/retryAskChatbot",
+  async (queryId, thunkAPI) => {
+    try {
+      const { session, query, courseId } = thunkAPI.getState().chatbot;
+
+      const queryItem = query.items.find((item) => item.id === queryId);
+
+      sendMessageToChatbot(queryItem.query_msg, session.id, courseId);
+
+    } catch (error) {
+      console.error(error);
+      return thunkAPI.rejectWithValue(_composeErrorMessage(error));
+    }
+  }
+);
+
+
+export const finishChatbotResponse = createAsyncThunk(
+  "chatbot/finishChatbotResponse",
+  async (errorMsg, thunkAPI) => {
+    try {
+      const { query } = thunkAPI.getState().chatbot;
+      const queryItem = query.items.find(
+        (item) => item.status === "writing" || item.status === "pending"
+      );
+
+      let status;
+
+      if (queryItem) {
+        status = errorMsg ? "failed" : "succeeded";
+        await updateQueryItemToDB({
+          id: queryItem.id,
+          response_msg: status === "succeeded" ? queryItem.response_msg : "",
+          status: status,
+        });
+      }
+
+      if (errorMsg) {
+        saveChabotErrorToDB(errorMsg)
+          .then(() => {
+            console.log("Saved chatbot error to db");
+          })
+          .catch(console.error);
+      }
+
+      /**
+       * 1. if status is undefined
+       *    this is an error not relevant to any writing item (no writing item)
+       *    so we don't need to update query list
+       *
+       * 2. otherwise we need to update writing query with corresponding status
+       */
+
+      return status;
+    } catch (error) {
+      /**
+       * error due to requesting to server
+       * that means there is an error with writing item
+       * => we need to update writing item status to failed in finishChatbotResponse.rejected
+       */
       console.error(error);
       return thunkAPI.rejectWithValue(_composeErrorMessage(error));
     }
@@ -58,35 +136,11 @@ export const fetchSessions = createAsyncThunk(
       const { session } = thunkAPI.getState().chatbot;
 
       const response = await fetchSessionsAPI(session.items.length, LIMIT);
-      console.log("session list response", response);
       return response.data;
     } catch (error) {
       console.error(error);
       return thunkAPI.rejectWithValue(_composeErrorMessage(error));
-    }
-  }
-);
 
-export const retryAskChatbot = createAsyncThunk(
-  "chatbot/retryAskChatbot",
-  async (queryId, thunkAPI) => {
-    try {
-      const response = await retryAskChatbotAPI(queryId);
-
-      return {
-        queryId: queryId,
-        response_msg: response.data.response_msg,
-        status: response.data.status,
-        error: "",
-      };
-    } catch (error) {
-      console.error(error);
-      return {
-        queryId: queryId,
-        response_msg: "",
-        status: "failed",
-        error: _composeErrorMessage(error),
-      };
     }
   }
 );
@@ -131,6 +185,9 @@ const initialState = {
     input: "",
     history: [],
     current: -1,
+
+    id: -1,
+
   },
 
   query: {
@@ -160,12 +217,27 @@ const initialState = {
     status: "idle",
     error: "",
   },
+
+
+  connection: {
+    status: "idle",
+    error: "",
+    retry: 0,
+  },
+
+  courseId: null,
+
 };
 
 const chatbotSlice = createSlice({
   name: "chatbot",
   initialState,
   reducers: {
+
+    reConnect: (state, action) => {
+      state.connection.retry += 1;
+    },
+
     setInputText: (state, action) => {
       state.ask.input = action.payload;
       state.ask.history = state.ask.history.map((item, index) => {
@@ -191,19 +263,29 @@ const chatbotSlice = createSlice({
       ];
     },
     startNewSession: (state, action) => {
-      state.query.items = [];
+
+
       state.session.id = 0;
     },
     changeSession: (state, action) => {
       state.session.id = action.payload;
-      state.query.items = [];
+
+      // state.query.items = [];
+
       state.query.isLastPage = false;
     },
     setRetryAskChatbotStatus: (state, action) => {
       state.query.items = state.query.items.map((item) =>
-        item.id === action.payload ? { ...item, status: "pending" } : item
+
+        item.id === action.payload
+          ? { ...item, status: "pending", response_msg: "" }
+          : item
       );
     },
+    setChatbotCourseId: (state, action) => {
+      state.courseId = action.payload || "NO_COURSE_ID";
+    },
+
     showChatbotFeedbackModal: (state, action) => {
       state.feedback.isShowModal = true;
       state.feedback.queryId = action.payload;
@@ -226,6 +308,38 @@ const chatbotSlice = createSlice({
       state.ask.input = state.ask.history[newIndex];
       state.ask.current = newIndex;
     },
+
+    writeChatbotResponse: (state, action) => {
+      if (!action.payload) return;
+
+      state.query.items = state.query.items.map((item) =>
+        item.status === "pending" || item.status === "writing"
+          ? {
+              ...item,
+              response_msg:
+                (item.response_msg || "") +
+                (action.payload === "<<Response Finished>>"
+                  ? ""
+                  : action.payload),
+              status:
+                action.payload !== "<<Response Finished>>"
+                  ? "writing"
+                  : "succeeded",
+            }
+          : item
+      );
+    },
+    connectionClose: (state, action) => {
+      state.connection.status = "failed";
+    },
+    connectionOpen: (state, action) => {
+      state.connection.error = "";
+      state.connection.status = "succeeded";
+    },
+    connectionError: (state, action) => {
+      state.connection.error = action.payload;
+    },
+
   },
   extraReducers: (builder) => {
     builder
@@ -242,6 +356,9 @@ const chatbotSlice = createSlice({
             vote: undefined,
             status: "pending",
             hash: action.meta.requestId,
+
+            session_id: state.session.id,
+
           },
         ];
 
@@ -256,33 +373,49 @@ const chatbotSlice = createSlice({
         state.ask.input = "";
       })
       .addCase(askChatbot.fulfilled, (state, action) => {
-        state.ask.status = "succeeded";
+
+
         state.query.items = state.query.items.map((item) =>
           item.hash === action.meta.requestId
             ? {
                 ...item,
-                response_msg: action.payload.response_msg,
-                status: action.payload.status,
+
                 id: action.payload.id,
+                created: action.payload.created,
+                session_id: action.payload.session_id,
+
               }
             : item
         );
 
         if (!state.query.initiated) state.query.initiated = true;
 
-        if (state.session.id !== 0) return;
 
-        state.session.id = action.payload.session_id;
-        state.session.items = [action.payload, ...state.session.items];
+        if (state.session.id === 0) {
+          state.session.id = action.payload.session_id;
+          state.session.items = [action.payload, ...state.session.items];
+        } else {
+          state.session.items = state.session.items.map((item) =>
+            item.session_id === action.payload.session_id
+              ? action.payload
+              : item
+          );
+        }
+
+        state.ask.id = action.payload.id;
+
       })
       .addCase(askChatbot.rejected, (state, action) => {
         state.ask.status = "failed";
         state.ask.error = action.payload;
         state.query.items = state.query.items.map((item) =>
           item.hash === action.meta.requestId
-            ? { ...item, status: "failed" }
+
+            ? { ...item, status: "failed", response_msg: "" }
             : item
         );
+        state.ask.id = -1;
+
       })
       // fetch sessions
       .addCase(fetchSessions.pending, (state, action) => {
@@ -290,7 +423,8 @@ const chatbotSlice = createSlice({
         state.session.error = "";
       })
       .addCase(fetchSessions.fulfilled, (state, action) => {
-        console.log("payload: ", action.payload);
+
+
         state.session.status = "succeeded";
         state.session.items = [
           ...state.session.items,
@@ -312,10 +446,21 @@ const chatbotSlice = createSlice({
       .addCase(fetchQueries.fulfilled, (state, action) => {
         state.query.status = "succeeded";
         action.payload.query_list.reverse();
-        state.query.items = [
-          ...action.payload.query_list,
+
+        // state.query.items = [
+        //   ...action.payload.query_list.map((item) =>
+        //     item.status === "idle" ? { ...item, status: "failed" } : item
+        //   ),
+        //   ...state.query.items,
+        // ];
+
+        state.query.items = _removeDuplicatedQueries([
+          ...action.payload.query_list.map((item) =>
+            item.status === "idle" ? { ...item, status: "failed" } : item
+          ),
           ...state.query.items,
-        ];
+        ]);
+
 
         if (action.payload.remain_page === 0) {
           state.query.isLastPage = true;
@@ -337,20 +482,27 @@ const chatbotSlice = createSlice({
       .addCase(retryAskChatbot.pending, (state, action) => {
         state.ask.status = "pending";
         state.ask.error = "";
-      })
-      .addCase(retryAskChatbot.fulfilled, (state, action) => {
-        state.ask.status = action.payload.status;
+
         state.query.items = state.query.items.map((item) =>
-          item.id === action.payload.queryId
-            ? {
-                ...item,
-                response_msg: action.payload.response_msg,
-                status: action.payload.status,
-              }
+          item.id === action.meta.arg
+            ? { ...item, status: "pending", response_msg: "" }
             : item
         );
-        state.ask.error = action.payload.error;
       })
+      // .addCase(retryAskChatbot.fulfilled, (state, action) => {
+      //   state.ask.status = action.payload.status;
+      //   state.query.items = state.query.items.map((item) =>
+      //     item.id === action.payload.queryId
+      //       ? {
+      //           ...item,
+      //           response_msg: action.payload.response_msg,
+      //           status: action.payload.status,
+      //         }
+      //       : item
+      //   );
+      //   state.ask.error = action.payload.error;
+      // })
+
       .addCase(retryAskChatbot.rejected, (state, action) => {
         //
       })
@@ -390,6 +542,56 @@ const chatbotSlice = createSlice({
       .addCase(giveFeedbackChatbot.rejected, (state, action) => {
         state.feedback.status = "failed";
         state.feedback.error = action.payload;
+
+      })
+      // finish
+      .addCase(finishChatbotResponse.fulfilled, (state, action) => {
+        if (!action.payload) {
+          return;
+        }
+
+        let updatedItem;
+        state.query.items = state.query.items.map((item) => {
+          if (item.status === "writing" || item.status === "pending") {
+            updatedItem = {
+              ...item,
+              status: action.payload,
+              response_msg:
+                action.payload === "succeeded" ? item.response_msg : "",
+            };
+
+            return updatedItem;
+          }
+          return item;
+        });
+        state.ask.status = action.payload;
+        state.session.items = state.session.items.map((item) =>
+          item.session_id === updatedItem.session_id ? updatedItem : item
+        );
+
+        state.ask.id = -1;
+        state.ask.status = "succeeded";
+      })
+      .addCase(finishChatbotResponse.rejected, (state, action) => {
+        if (state.ask.status !== "pending") return;
+
+        let updatedItem;
+        state.query.items = state.query.items.map((item) => {
+          if (item.status === "writing" || item.status === "pending") {
+            updatedItem = { ...item, status: "failed", response_msg: "" };
+            return updatedItem;
+          }
+
+          return item;
+        });
+        state.ask.status = "failed";
+
+        state.session.items = state.session.items.map((item) =>
+          item.session_id === updatedItem.session_id ? updatedItem : item
+        );
+
+        state.ask.id = -1;
+
       });
   },
 });
@@ -404,6 +606,14 @@ export const {
   showChatbotFeedbackModal,
   hideChatbotFeedbackModal,
   setChatbotInputHistory,
+
+  writeChatbotResponse,
+  connectionClose,
+  connectionOpen,
+  reConnect,
+  setChatbotCourseId,
+  connectionError,
+
 } = chatbotSlice.actions;
 export default chatbotSlice.reducer;
 
@@ -421,3 +631,17 @@ function _composeErrorMessage(error) {
 
   return `${statusCode} - ${errorMsg}`;
 }
+
+
+function _removeDuplicatedQueries(queries) {
+  const result = [];
+
+  queries.forEach((query) => {
+    if (!result.find((item) => item.id === query.id)) {
+      result.push(query);
+    }
+  });
+
+  return result;
+}
+
